Add service for getting all bibliographicrecordids for an agency
An endpoint for getting all ids for an agency and an endpoint for
getting a list of records are needed for a new machine learning
application.

Note that the result returned by the bibliographicrecordids endpoint can
be very large. 120k+ records result in an 18 MB response. Next step is
to add compression.

The bibliographicrecordids endpoint returns a response that has the
same format as the request to the records/bulk endpoint, however sending
the whole list in one request is not advised!

If the bibliographicrecordids endpoint is called with a DBC agency the
result is enriched with the 191919 agency.