Add option to exclude the content attribute
The base64 encoded content field usually takes up more than half size of
the response, and if you are not going to need the content field it is
a waste of memory and unmarshalling time for the receiver.

Therefor an exclude-attribute URL param has been implemented. The reason
why exclude-attribute is not called exclude-field is because a field
in this context is a data field inside the content.

Note that only 'content' and 'contentJSON' attributes are supported
right now.

Also exclude-attribute is repeatable, e.g.
...&exclude-attribute=content&exclude-attribute=contentjson