Add record relations functions
The basis solr indexer needs the /siblings-from endpoint now and will
soon also need the /children endpoint. So I decided to implement the
other relations functions as well.

Right now it is kinda a mess in regards to how the rawrepo access DAO is
called. At some point the access DAO will be folded into this project
and then should look at how we can refactor the code.