Retrieve merged deleted records
Since relations are removed when a record is marked as deleted we have
to handle deleted records a bit different.

Expansion of deleted records are not supported at this time.

MarcRecordBean was moved to dk.dbc.rawrepo in order to gain
access to the package protected constructor function on RecordImpl.