Insert record service URL in configurations table on startup
This change is DIT. Some service rely on the URL to record service to be
in the configurations table. However in DIT where the database is
created from scratch on each restart there is no good way to initialize
the database with the record service URL. So now the service inserts its
own URL in the table.

WARNING This can have adverse affect if used in production!