Don't throw away the requested authority record
If you requested a record collection with exclude-aut-records=true and
the record you requested the collection from is a authority record then
you would get an empty collection back.

In order to be more compatible with the behavior of rawrepo-dao this
situation has been handled by always including the requested record even
if that record is an authority record.