Fix indicators
There were two problems:
First off indicator 2 was included twice and indicator 3 was missing.
Now all three indicators are actually read.

Secondly the indicator field should only be two chars long when
converting from marcxchange, so now indicator 3 is excluded in the
output if the indicator is null. And indicator 1 and indicator 2 are
replaced by space if they are missing.