Catch exceptions during thread execution
invokeAll absorbs all exceptions which makes it difficult to report back
any errors that might have happened during execution. invokeAny will
throw an ExecutionException if a thread for some reason doesn't report
OK back. This allows us to catch ExecutionException and rethrow it to
something we can report back to the user.