Report the number of records correctly
The problem was that both local and enrichment record was already
counted regardless of which record types was actually chosen.

E.g. there are 50 local records and 100 enrichments for an FBS library.
If dump is performed with -t LOCAL the record service would return 150
as the amount of records but the dump will only contain 50 records.

Now only local record count is returned if LOCAL is chosen and vice
versa for enrichments.

In addition it turns out that invokeAny doesn't invoke all the
threads... It seems like it stops halfway or when the first set of
threads has looped once. Anyway the result was not all records was
written to the file. A solution which both returns a proper
MarcReaderException exception and doesn't stop halfway is to use
invokeAll and then check afterwards if any threads threw an exception.