Paginate how many holdingsitems are checked against rawrepo
Well it turns out that some libraries have so many holdings items that
they cannot be checked against rawrepo in a single query. Instead we
have to page over the holdings items and check a slice at the time.

This can be quite time consuming but it probably won't be used that
often. If you dump holdings for a library with a lot of holdings then
you most likely already know that. When dumping DBC or local libraries
or FBS without holdings this won't have any performance impact.