Only do exact match during dry run
Dryrun is used for outputting the exact record count to the user. But
the dump don't need the exact count as it will just skip those records
in rawrepo.

So in order to save some unnecessary queries to rawrepo we will skip the
exact look up when just dumping records.